+++ src/OpenSearch/Common/Exceptions/Conflict409Exception.php
@@ -21,6 +21,11 @@ declare(strict_types=1);
 
 namespace OpenSearch\Common\Exceptions;
 
+@trigger_error(Conflict409Exception::class . ' is deprecated in 2.3.2 and will be removed in 3.0.0.', E_USER_DEPRECATED);
+
+/**
+ * @deprecated in 2.3.2 and will be removed in 3.0.0.
+ */
 class Conflict409Exception extends \Exception implements OpenSearchException
 {
 }