+++ src/OpenSearch/Common/Exceptions/MaxRetriesException.php
@@ -21,6 +21,11 @@ declare(strict_types=1);
 
 namespace OpenSearch\Common\Exceptions;
 
+@trigger_error(MaxRetriesException::class . ' is deprecated in 2.3.2 and will be removed in 3.0.0.', E_USER_DEPRECATED);
+
+/**
+ * @deprecated in 2.3.2 and will be removed in 3.0.0.
+ */
 class MaxRetriesException extends TransportException implements OpenSearchException
 {
 }