+++ src/OpenSearch/ConnectionPool/Selectors/RoundRobinSelector.php
@@ -23,6 +23,11 @@ namespace OpenSearch\ConnectionPool\Selectors;
 
 use OpenSearch\Connections\ConnectionInterface;
 
+@trigger_error(RoundRobinSelector::class . ' is deprecated in 2.3.2 and will be removed in 3.0.0.', E_USER_DEPRECATED);
+
+/**
+ * @deprecated in 2.3.2 and will be removed in 3.0.0.
+ */
 class RoundRobinSelector implements SelectorInterface
 {
     /**