+++ src/OpenSearch/ConnectionPool/SimpleConnectionPool.php
@@ -22,10 +22,14 @@ declare(strict_types=1);
 namespace OpenSearch\ConnectionPool;
 
 use OpenSearch\ConnectionPool\Selectors\SelectorInterface;
-use OpenSearch\Connections\Connection;
 use OpenSearch\Connections\ConnectionFactoryInterface;
 use OpenSearch\Connections\ConnectionInterface;
 
+@trigger_error(SimpleConnectionPool::class . ' is deprecated in 2.3.2 and will be removed in 3.0.0.', E_USER_DEPRECATED);
+
+/**
+ * @deprecated in 2.3.2 and will be removed in 3.0.0.
+ */
 class SimpleConnectionPool extends AbstractConnectionPool implements ConnectionPoolInterface
 {
     /**